+++ plugin/sampling/internal/leaderelection/mocks/ElectionParticipant.go
